Move updateQueryContext to ExploreChartPanel
